sample-app/agent/pyproject.toml
[project]
name = "prism-d1-task-agent"
version = "0.1.0"
description = "PRISM D1 Velocity — Task Assistant Agent (Strands SDK + MCP + AgentCore)"
requires-python = ">=3.11"
dependencies = [
    "strands-agents>=1.0.0",
    "strands-agents-tools>=0.4.0",
    "boto3>=1.34.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "moto[bedrock]>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
